app/requirements.txt
spotipy
python-dotenv
pandas
Flask
Flask-Cors
opendatasets
matplotlib
scikit-learn